requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
markdown>=3.5.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.2.0
flask>=3.0.0
gunicorn>=21.2.0
